O cor dos botões estão mudando, quando like ou deslike
