+++ modulo3/projeto-labeddit/projeto-lab-eddit/src/components/ContadorVotosComentarios/styled.js
@@ -1,0 +1,54 @@
+import styled from "styled-components"
+
+export const BotaoLike = styled.button`
+width: 1.8vw;
+height : 3.5vh;
+font-size: 1em;
+border-radius: 3px;
+font-size: 15px;
+border: 1px solid blue;
+color: blue;
+background-color: white;
+&:hover{
+    color: white;
+    background-color: blue;
+    }
+`
+
+export const BotaoLike2 = styled.button`
+width: 1.8vw;
+height : 3.5vh;
+font-size: 1em;
+border-radius: 3px;
+font-size: 15px;
+border: 1px solid blue;
+color: white;
+background-color: blue;
+`
+
+export const BotaoDeslike = styled.button`
+width: 1.8vw;
+height : 3.5vh;
+font-size: 1em;
+border-radius: 3px;
+font-size: 15px;
+border: 1px solid red;
+color: red;
+background-color: white;
+margin-right: 0.5vw;
+&:hover{
+    color: white;
+    background-color: red;
+    }
+`
+
+export const BotaoDeslike2 = styled.button`
+width: 1.8vw;
+height : 3.5vh;
+font-size: 1em;
+border-radius: 3px;
+font-size: 15px;
+border: 1px solid red;
+color: white;
+background-color: red;
+`
