+++ modulo3/projeto-labeddit/projeto-lab-eddit/src/pages/FeedPage/styled.js
@@ -102,56 +102,3 @@ span{
     }
 }
 `
-
-export const BotaoLike = styled.button`
-width: 1.8vw;
-height : 3.5vh;
-font-size: 1em;
-border-radius: 3px;
-font-size: 15px;
-border: 1px solid blue;
-color: blue;
-background-color: white;
-&:hover{
-    color: white;
-    background-color: blue;
-    }
-`
-
-export const BotaoLike2 = styled.button`
-width: 1.8vw;
-height : 3.5vh;
-font-size: 1em;
-border-radius: 3px;
-font-size: 15px;
-border: 1px solid blue;
-color: white;
-background-color: blue;
-`
-
-export const BotaoDeslike = styled.button`
-width: 1.8vw;
-height : 3.5vh;
-font-size: 1em;
-border-radius: 3px;
-font-size: 15px;
-border: 1px solid red;
-color: red;
-background-color: white;
-margin-right: 0.5vw;
-&:hover{
-    color: white;
-    background-color: red;
-    }
-`
-
-export const BotaoDeslike2 = styled.button`
-width: 1.8vw;
-height : 3.5vh;
-font-size: 1em;
-border-radius: 3px;
-font-size: 15px;
-border: 1px solid red;
-color: white;
-background-color: red;
-`
