+++ modulo3/projeto-labeddit/projeto-lab-eddit/src/pages/PostPage/PostPage.js
@@ -1,48 +1,20 @@
-import React, { useState } from "react"
-import { Posts, Map, Comment, BotaoLike, BotaoDeslike, BotaoVoltar } from "./styled"
+import React from "react"
+import { Posts, Map, Comment, BotaoVoltar } from "./styled"
 import useForm from "../../hooks/useForm"
 import useProtectedPage from "../../hooks/useProtectedPage"
-import { createComment, useData, createCommentVote, changeCommentVote, deleteCommentVote, useResquestData } from "../../requests/requests"
+import { createComment, useData } from "../../requests/requests"
 import { goToFeed } from "../../router/coordinatis"
 import { useHistory, useParams } from "react-router-dom"
 import Headers from "../../components/Headers/Headers"
+import ContadorVotosPost from "../../components/ContadorVotoPost/ContadorVotosPost"
+import ContadorComentarios from "../../components/ContadorVotosComentarios/ContadorComentarios"
+import { useResquestData } from "../../hooks/useResquestData"
 
 function PostPage() {
     useProtectedPage()
 
     const posts = useResquestData([])
     const history = useHistory()
-
-    //-------------- VOTAR EM UM COMENTÁRIO ----------
-    const [like, setLike] = useState(false)
-    const [deslike, setDeslike] = useState(false)
-    const vote = (id, valor) => {
-        const body = {
-            direction: valor
-        }
-
-        if (valor === 1 && !like && !deslike) {
-            setLike(true)
-            createCommentVote(id, body, valor)
-        } else if (valor === -1 && !like && !deslike) {
-            setDeslike(true)
-            createCommentVote(id, body, valor)
-        } else if (valor === 1 && like) {
-            setLike(false)
-            deleteCommentVote(id)
-        } else if (valor === -1 && deslike) {
-            setDeslike(false)
-            deleteCommentVote(id)
-        } else if (valor === 1 && deslike) {
-            setLike(true)
-            setDeslike(false)
-            changeCommentVote(id, body, valor)
-        } else {
-            setLike(false)
-            setDeslike(true)
-            changeCommentVote(id, body, valor)
-        }
-    }
 
     // --------------- FUNÇÕES DO FORMULARIO----------
     const { form, onChange, cleanFields } = useForm({
@@ -66,7 +38,13 @@ function PostPage() {
                 <h2>{post.title}</h2>
                 <p>{post.body}</p>
                 <span>
-                    <p> <b>votos:</b> {post.voteSum ? post.voteSum : 0} </p>
+                    <div>
+                        <ContadorVotosPost
+                            id={post.id}
+                            votos={post.voteSum}
+                        />
+                    </div>
+                    <div>|</div>
                     <p>{post.commentCount ? post.commentCount : 0} 💬 </p>
                 </span>
             </Map>
@@ -82,9 +60,10 @@ function PostPage() {
                 <span>
                     <p>{item.body}</p>
                     <div>
-                        <BotaoLike onClick={() => { vote(item.id, 1) }} > <b>↑</b> </BotaoLike>
-                        <p>{item.voteSum ? item.voteSum : 0}</p>
-                        <BotaoDeslike onClick={() => { vote(item.id, -1) }}><b>↓</b></BotaoDeslike>
+                        <ContadorComentarios
+                            id={item.id}
+                            votos={item.voteSum}
+                        />
                     </div>
 
                 </span>
